Redirect user to login page after logout

+++ client/src/components/Sidebar.js
@@ -12,6 +12,7 @@ import SearchUser from './SearchUser';
 import { FaImage } from "react-icons/fa6";
 import { GoVideo } from "react-icons/go";
 import { logout } from '../redux/userSlice';
+import axios from 'axios';
 
 
 
@@ -54,12 +55,27 @@ const Sidebar = () => {
     }
   }, [socketConnection, user]);
 
-  const handleLogout= ()=>{
-    dispatch(logout())
-    navigate("/email")
-    localStorage.clear()
+  const handleLogout= async()=>{
+    try{
+      
+      let response= await axios.post(`${process.env.REACT_APP_BACKEND_URL}/api/logout`, {}, {withCredentials: true})
+      if(response.data.success){
+        alert(response.data.message)
+        dispatch(logout())
+        localStorage.clear()
+        navigate("/email")
+      }
+    }
+    catch(err){
+      console.log(err)
+    }
   }
   
+    useEffect(() => {
+      if (!user._id) {
+        navigate("/email");
+      }
+    }, [user, navigate]);
 
 
   return (
